+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Only running React 16 tests on CircleCI (when an explicit envvar is set)
 -   Adds an example app with Email Verification with OTP
 
+### Changes
+
+-   Clearing errors when switching between sign-in and up
+
 ## [0.22.0] - 2022-06-03
 
 -   Adds a SuperTokens + Supabase example app